Health-kiosk/kiosk_be/requirements.txt
Adafruit-Blinka==6.9.2
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-mlx90614==1.2.6
adafruit-extended-bus==1.0.1
Adafruit-PlatformDetect==3.13.2
Adafruit-PureIO==1.1.8
astroid==2.5.2
board==1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
datedelta==1.3
DateTime==4.3
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.2
mccabe==0.6.1
mysql-connector-python==8.0.23
protobuf==3.15.7
pycparser==2.20
pyftdi==0.53.1
PyJWT==2.0.1
pylint==2.7.4
PyMySQL==1.0.2
pyserial==3.5
python-dotenv==0.17.0
pytz==2021.1
pyusb==1.1.1
six==1.15.0
SQLAlchemy==1.4.5
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.3.0
